Fix package-analysis.json inconsistency issue.
`package-analysis.json` is generated by `GenerateJsonReport`. In this
method, `Task.ContinueWith` is used to filter the analysis results, `Task.ContinueWith` creates a continuation asynchronous Task that executes when its antecedent has completed. And these continuation tasks are not waited to finish before the program writes to json file. It's unnecessary to use `ContinueWith` for filtering the list, removing it fixes the inconsistency issue.