Fix integration test failure.
The integration test project refers to application project (not library)
PortingAssistant.Client.CLI. In this case,
PortingAssistant.Client.CLI.runtimeconfig.json won't be copied to the
integration test project's output folder and we  won't be able to run
PortingAssistant.Client.CLI.exe binary because it will throw The library
hostpolicy.dll was not found.

This patch address the issue by copying runtimeconfig.json and
depend.json to the project that refers it.