Fix ThrowAddingDuplicateWithKeyArgumentException.
ApplyPortingChanges does an aggregation on the RecommendedActions across all projects being ported based on the package ID and generate a dictionary. When more than one project in a solution depend on a same package, the aggregation throws  ThrowAddingDuplicateWithKeyArgumentException due to duplicated key with the same package ID. This patch fix the issue by apply GroupBy before converting to dictionary.