Update Application_GUID generation.
With current change, the Applicatoin_GUID is calculated based on concatenation of all projectGuids instead of ProjectNames, this garantees the Application_GUID to be unique across customers when they have projects that have same names.

This patch also writes Solution_GUID, Application_GUID as well as Repository_Url to assessment json output.

Integration test are also updated to reflect this change.